move buttonviewmanager to shadownode
