+++ vnext/ReactUWP/Polyester/ButtonViewManager.cpp
@@ -17,6 +17,31 @@ namespace react {
 namespace uwp {
 namespace polyester {
 
+class ButtonShadowNode : public ContentControlShadowNode {
+  using Super = ContentControlShadowNode;
+
+ public:
+  ButtonShadowNode() = default;
+  void createView() override;
+
+ private:
+  winrt::Button::Click_revoker m_buttonClickRevoker{};
+};
+
+void ButtonShadowNode::createView() {
+  Super::createView();
+
+  auto button = GetView().as<winrt::Button>();
+
+  m_buttonClickRevoker =
+      button.Click(winrt::auto_revoke, [=](auto &&, auto &&) {
+        auto instance = GetViewManager()->GetReactInstance().lock();
+        folly::dynamic eventData = folly::dynamic::object("target", m_tag);
+        if (instance != nullptr)
+          instance->DispatchEvent(m_tag, "topClick", std::move(eventData));
+      });
+}
+
 ButtonViewManager::ButtonViewManager(
     const std::shared_ptr<IReactInstance> &reactInstance)
     : ContentControlViewManager(reactInstance) {}
@@ -44,16 +69,12 @@ folly::dynamic ButtonViewManager::GetExportedCustomDirectEventTypeConstants()
   return directEvents;
 }
 
+facebook::react::ShadowNode *ButtonViewManager::createShadow() const {
+  return new ButtonShadowNode();
+}
+
 XamlView ButtonViewManager::CreateViewCore(int64_t tag) {
   winrt::Button button = winrt::Button();
-  m_buttonClickRevoker =
-      button.Click(winrt::auto_revoke, [=](auto &&, auto &&) {
-        auto instance = m_wkReactInstance.lock();
-        folly::dynamic eventData = folly::dynamic::object("target", tag);
-        if (instance != nullptr)
-          instance->DispatchEvent(tag, "topClick", std::move(eventData));
-      });
-
   return button;
 }
