+++ vnext/ReactUWP/Polyester/ContentControlViewManager.h
@@ -4,10 +4,26 @@
 #pragma once
 
 #include <Views/ControlViewManager.h>
+#include <Views/ShadowNodeBase.h>
 
 namespace react {
 namespace uwp {
 namespace polyester {
+
+class ContentControlShadowNode : public ShadowNodeBase {
+  using Super = ShadowNodeBase;
+
+ public:
+  ContentControlShadowNode() = default;
+  void createView() override;
+  bool IsExternalLayoutDirty() const override {
+    return m_paddingDirty;
+  }
+  void DoExtraLayoutPrep(YGNodeRef yogaNode) override;
+
+ private:
+  bool m_paddingDirty = false;
+};
 
 class ContentControlViewManager : public ControlViewManager {
   using Super = ControlViewManager;
